chatbot_service/requirements.txt
django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
requests==2.31.0
celery==5.3.6
redis==5.0.1
tensorflow
numpy
markdown==3.5.1
gunicorn
python-dotenv
django-cors-headers